refactor: migrate to pyproject.toml with granular global_utils extras
