--- backend/requirements.txt
@@ -1,9 +1,0 @@
-pydantic
-pydantic-settings
-gunicorn
-flask
-flask-cors
-pymongo
-python-dotenv
-webargs
-requests
