+++ global_utils/pyproject.toml
@@ -1,0 +1,38 @@
+[build-system]
+requires = ["setuptools>=68.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "global_utils"
+version = "1.0.0"
+description = "Global shared utilities for UnifAI services"
+requires-python = ">=3.11"
+license = {text = "Proprietary"}
+
+dependencies = [
+    "pydantic>=2.0",
+    "pydantic-settings>=2.0",
+    "pyyaml",
+    "python-dotenv",
+]
+
+[project.optional-dependencies]
+flask           = ["flask"]
+apiargs         = ["webargs", "marshmallow"]
+redis           = ["redis"]
+celery          = ["celery"]
+docling         = ["httpx>=0.25.0"]
+embedding       = ["httpx>=0.25.0"]
+async-bridge    = ["anyio"]
+codegen         = ["datamodel-code-generator", "jsonschema"]
+
+rag             = ["global_utils[flask,apiargs,celery,docling,embedding]"]
+backend         = ["global_utils[flask,apiargs,redis]"]
+multi-agent     = ["global_utils[flask,apiargs,async-bridge,codegen]"]
+identity        = ["global_utils[flask,redis]"]
+
+all             = ["global_utils[flask,apiargs,redis,celery,docling,embedding,async-bridge,codegen]"]
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["global_utils*"]
