--- global_utils/requirements.txt
@@ -1,14 +1,0 @@
-webargs
-marshmallow
-celery
-celery-batches
-flask_pymongo
-pymongo
-asyncio
-datamodel-code-generator
-pydantic
-pydantic-settings
-jsonschema
-httpx>=0.25.0
-anyio
-redis
