--- multi-agent/requirements.txt
@@ -1,27 +1,0 @@
-pydantic
-pydantic-core
-pyyaml
-typer
-langchain
-langchain-core
-langchain-openai
-langchain-google-genai
-openai
-langgraph
-gunicorn
-flask
-flask-cors
-pymongo
-python-dotenv
-httptools
-mcp
-paramiko
-openshift-client
-yarl
-jsonschema
-html-to-markdown
-safehttpx
-authlib
-anyio
-typing-inspection
-a2a-sdk[all]<1.0
