+++ rag/core/data_sources/types/document/pipeline_handler.py
@@ -10,7 +10,7 @@ from infrastructure.sources.document.connector import DocumentConnector
 from infrastructure.sources.document.chunker import PDFChunkerStrategy
 from shared.logger import logger
 
-from global_utils.utils import cleanup_file
+from global_utils.utils.file_utils import cleanup_file
 
 
 class DocumentPipelineHandler(SourcePipelinePort):
