--- rag/requirements.txt
@@ -1,17 +1,0 @@
-flask 
-flask-cors
-requests
-tiktoken
-langchain==0.3.25
-langchain-core==0.3.65
-langchain-openai==0.3.24
-numpy
-# sentence-transformers
-qdrant-client
-# docling
-Authlib
-PyJWT
-python-dotenv
-pydantic==2.11.7
-umami-analytics
-pymongo
