--- shared-resources/identity/requirements.txt
@@ -1,9 +1,0 @@
-flask
-flask-cors
-requests
-Authlib
-PyJWT
-python-dotenv
-pydantic
-pydantic-settings
-redis
